[docs] Remove unused abstraction
